requirements.txt
behave~=1.2.6
selenium
webdriver-manager
pyyaml
PyHamcrest